all error response format and static method implentation done

+++ src/app/modules/user/user.service.ts
@@ -59,8 +59,11 @@ const UpdateUserFromDB = async (userId: number, user: IUser) => {
 const deleteUserFromDB = async (userId: number) => {
 
     try {
-        const result = await User.deleteOne({ userId });
-        return result;
+        const res = await User.isUserExists(userId);
+        if (res) {
+            await User.deleteOne({ userId });
+        }
+        return res;
     } catch (error) {
         console.error('Error saving user:', error);
         throw error; // Re-throw the error to handle it at a higher level if needed
